Fix bug in HidatoCSP
There was a bug which caused non-square boards to be unsolvable.
The bug was in the method _in_grid: x, y was interperted as col, row
instead of the col, row interpertation used throughout the class,
so x was compared to width and y was compared to height instead of
vice-versa which is the correct way.