Change the way lookup is created in get_grid
get_grid was the function with the longest comulative runtime.
The most time-consuming line was the line creating the lookup dict by
passing a generator expression to the dict constructor.
By constructing the dict as a generator expression explicitly,
runtime was reduced by a very large factor, enabling the creation of
10 x 10 boards in just a few seconds compared to minutes.